+++ src/nodejs/lambda-handlers/submission.js
@@ -470,7 +470,7 @@ async function deleteStepReviewApprovalMethod(event, user) {
 }
 
 async function assignDaacsMethod(event, user) {
-  const { id, daacs } = event;
+  const { id, daacs, requires_review: requiresReview } = event;
   const approvedUserPrivileges = ['ADMIN', 'REQUEST_ASSIGNDAAC'];
 
   if (!user.user_privileges.some((privilege) => approvedUserPrivileges.includes(privilege))) {
@@ -479,62 +479,106 @@ async function assignDaacsMethod(event, user) {
 
   let submission = await db.submission.findById({ id, user_id: user.id });
 
+  if (submission.error) {
+    return { error: 'Invalid permissions.' };
+  }
+
   // Check current step - only proceed if on DAAC assignment step
-  if (submission.step_name !== 'daac_assignment') {
+  if (!submission.step_name.match(/daac_assignment(_final)?/g)) {
     return { error: 'Invalid workflow step. Unable to assign DAACs.' };
   }
 
-  // Generate a code for each daac to be assigned
-  // eslint-disable-next-line
-  for (const daacId of daacs) {
-    await db.submission.createCode({ submissionId: id, daacID: daacId });
-  }
-
-  submission = await db.submission.findById({ id, user_id: user.id });
-
-  // Send notification emails to the following
-  // - the point of contact of the submission (form response),
-  // - the assigned DAAC's Data Managers,
-  // - and the ESDIS observer collaborator of the accession request
-
-  const idList = [];
-
-  // Get the assigned DAAC's Data Managers Ids
-  // eslint-disable-next-line
-  for (const assignedDaac of submission.assigned_daacs) {
-    const managerList = await db.user.getManagerIds({ daac_id: assignedDaac.daac_id });
-
-    if (Array.isArray(managerList)) {
-      managerList.forEach((entry) => idList.push(entry.id));
+  // If the submission requires a daac review - assign the daac and move to the next step
+  if (requiresReview) {
+    // Double check that only one DAAC was chosen
+    // should not be an issue comming from the dashboard but ensure is the case from the API
+    if (daacs && daacs.length > 1) {
+      return { error: 'Invalid number of DAACS selected. Only 1 DAAC may be selected when needs review is true.' };
     }
+
+    // Assign Daac to the submission
+    const daacId = daacs[0];
+    await db.submission.updateDaac({ id, daac_id: daacId });
+  } else {
+    // If the submission does not require a daac review - assign the codes for all daacs selected
+    // If there are existing codes - remove any not in the new list
+    if (submission.assigned_daacs !== null) {
+      const removedDaacs = submission.assigned_daacs
+        .filter((daacObj) => !daacs.includes(daacObj.daac_id))
+        .map((daacObj) => daacObj.daac_id);
+      await db.submission.deleteCodes({ submissionId: id, daacs: removedDaacs });
+    }
+
+    // Generate a code for each daac to be assigned
+    // eslint-disable-next-line
+    for (const daacId of daacs) {
+      await db.submission.createCode({ submissionId: id, daacID: daacId });
+    }
+
+    submission = await db.submission.findById({ id, user_id: user.id });
+
+    // Send notification emails to the following
+    // - the point of contact of the submission (form response),
+    // - the assigned DAAC's Data Managers,
+    // - and the ESDIS observer collaborator of the accession request
+
+    const userIds = [];
+
+    // Get the assigned DAAC's Data Managers Ids
+    // eslint-disable-next-line
+    for (const assignedDaac of submission.assigned_daacs) {
+      const managerList = await db.user.getManagerIds({ daac_id: assignedDaac.daac_id });
+
+      if (Array.isArray(managerList)) {
+        managerList.forEach((entry) => userIds.push(entry.id));
+      }
+    }
+
+    // Get the ESDIS observer collaborator of the accession request
+    const rootOnly = true;
+    const observers = await db.user.getObserverIds({
+      contributor_ids: submission.contributor_ids,
+      root_only: rootOnly
+    });
+
+    if (Array.isArray(observers)) {
+      observers.forEach((entry) => userIds.push(entry.id));
+    }
+
+    // TODO - Revisit this if we start autopopulating data from the first form to the second
+    const pocRecipients = [];
+
+    // Add the POC from the first form
+    if (submission.form_data.assignment_form_data_submission_poc_email) {
+      pocRecipients.push({
+        name: submission.form_data.assignment_form_data_submission_poc_name
+          ? submission.form_data.assignment_form_data_submission_poc_name : '',
+        email: submission.form_data.assignment_form_data_submission_poc_email
+      });
+    }
+
+    // Add the POC from the second form
+    if (submission.form_data.poc_email) {
+      pocRecipients.push({
+        name: submission.form_data.poc_name ? submission.form_data.poc_name : '',
+        email: submission.form_data.poc_email
+      });
+    }
+
+    const eventMessage = {
+      event_type: 'daac_assignment',
+      submission_id: submission.id,
+      conversation_id: submission.conversation_id,
+      submission_name: submission.name,
+      step_name: submission.step_name,
+      assigned_daacs: submission.assigned_daacs,
+      ...(userIds.length > 0 && { userIds }),
+      ...(pocRecipients.length > 0 && { additional_recipients: pocRecipients }),
+      emailPayloadProvided: 'true'
+    };
+
+    await msg.sendEvent(eventMessage);
   }
-
-  // Get the ESDIS observer collaborator of the accession request
-  const observers = await db.user.getObserverIds({ contributor_ids: submission.contributor_ids });
-
-  if (Array.isArray(observers)) {
-    observers.forEach((entry) => idList.push(entry.id));
-  }
-
-  const emailRecipients = await db.user.getEmails({ user_list: idList });
-
-  if (!(emailRecipients && Array.isArray(emailRecipients))) {
-    return { error: 'Invalid Recipients. Code notification emails were not sent.' };
-  }
-
-  // Add the point of contact info from the submission
-  emailRecipients.push({
-    name: submission.form_data.poc_name,
-    email: submission.form_data.poc_email
-  });
-
-  const emailPayload = {
-    event_type: 'daac_assignment',
-    submission_name: submission.name,
-    assigned_daacs: submission.assigned_daacs
-  };
-
-  await msg.sendEmail(emailRecipients, emailPayload);
 
   // Promote to the next workflow step
   const eventMessage = {
@@ -551,6 +595,76 @@ async function assignDaacsMethod(event, user) {
   return db.submission.findById({ id, user_id: user.id });
 }
 
+async function esdisReviewMethod(event, user) {
+  const approvedUserPrivileges = ['ADMIN'];
+  const conditionalUserPrivileges = ['REQUEST_REVIEW_ESDIS'];
+  const validWorkflowSteps = ['esdis_final_review'];
+  const { id, action } = event;
+  if (
+    user.id?.includes('service-authorizer')
+    || user.user_privileges.some((privilege) => approvedUserPrivileges.includes(privilege))
+    || (user.user_privileges.some((privilege) => conditionalUserPrivileges.includes(privilege))
+    && user.user_groups.some((group) => group.short_name === 'root_group'))
+  ) {
+    const status = await db.submission.getState({ id });
+    if (!validWorkflowSteps.includes(status.step_name)) {
+      return { error: 'Invalid workflow step. Unable to complete review.' };
+    }
+    let eventType;
+    let param;
+    let nextStep;
+    if (action === 'reject') {
+      eventType = 'review_rejected';
+      nextStep = 'close';
+    } else if (action === 'approve') {
+      eventType = 'review_approved';
+    } else if (action === 'reassign') {
+      eventType = 'workflow_step_change';
+
+      // reset any current review requirements (assignees) for the DAR review step
+      const currentReviewers = await db.submission.getStepReviewApproval({ id });
+      const userIds = currentReviewers.map((reviewer) => reviewer.edpuser_id);
+      const deleteReviewStep = 'data_accession_request_form_review';
+
+      // eslint-disable-next-line
+      for (const userId of userIds) {
+        const eventData = {
+          submissionId: id,
+          userIds: [userId],
+          stepName: deleteReviewStep
+        };
+        await deleteStepReviewApprovalMethod(eventData, user);
+      }
+
+      // remove the currently assigned DAAC value
+      await db.submission.updateDaac({ id, daac_id: null });
+
+      // reset the current step to the DAAC Assignment Prompt step
+      param = {
+        id,
+        step_name: 'daac_assignment'
+      };
+      await changeStepMethod(param, user);
+    } else {
+      return { error: 'Invalid review action' };
+    }
+    const eventMessage = {
+      event_type: eventType,
+      submission_id: id,
+      conversation_id: status.conversation_id,
+      workflow_id: status.workflow_id,
+      user_id: user.id,
+      data: status.step.data,
+      step_name: status.step_name,
+      ...(nextStep && { next_step: nextStep })
+    };
+    if (status.step.step_message) eventMessage.step_message = status.step.step_message;
+    await msg.sendEvent(eventMessage);
+    return db.submission.getState({ id });
+  }
+  return { error: 'Not Authorized' };
+}
+
 const operations = {
   initialize: initializeMethod,
   active: statusMethod,
@@ -560,6 +674,7 @@ const operations = {
   submit: submitMethod,
   save: saveMethod,
   review: reviewMethod,
+  esdisReview: esdisReviewMethod,
   resume: resumeMethod,
   lock: lockMethod,
   unlock: unlockMethod,