+++ src/nodejs/lambda-layers/database-util/src/query/parsers.js
@@ -125,6 +125,7 @@ module.exports.stepCleanup = one;
 module.exports.create = one;
 module.exports.checkCode = one;
 module.exports.createCode = one;
+module.exports.deleteCodes = one;
 module.exports.createSection = one;
 module.exports.addAttachments = one;
 module.exports.findAllUploadSteps = many;