+++ src/nodejs/lambda-layers/message-util/src/message-util.js
@@ -75,6 +75,8 @@ function marshalAttribute(attribute) {
   } if (typeof attribute === 'number') {
     return { DataType: 'Number', StringValue: `${attribute}` };
   }
+  // eslint-disable-next-line no-console
+  console.log('marshalAttribute: Unsupported attribute type', attribute);
   return {};
 }