+++ src/postgres/7-seed.sql
@@ -3,6 +3,7 @@
 INSERT INTO form VALUES ('6c544723-241c-4896-a38c-adbc0a364293', 'data_accession_request', 1, 'Data Accession Request', 'This form is used to get high level information about a dataset, typically this will be submitted by the data provider or an appropriate agent.');
 INSERT INTO form VALUES ('19025579-99ca-4344-8610-704dae626343', 'data_publication_request', 1, 'Data Publication Request', 'This form is used to get high level information about a dataset, typically this will be submitted by the data provider or an appropriate agent.');
 INSERT INTO form VALUES ('3f77385f-7087-4d22-81c1-5c29b95d3295', 'ornl_submission', 1, 'ORNL Submission Form', 'This form is used by ORNL DAAC staff to enter pieces of information needed by the DAAC that the Data Provider will not have when submitting the Data Accession Request Form', true);
+INSERT INTO form VALUES ('19025579-99ca-4344-8611-704dae626343', 'Submit_Data_ESDIS_DAAC', 1, 'Submit Data to an ESDIS DAAC', 'Submit Data to an ESDIS DAAC.');
 
 -- Service(id, short_name, long_name, description, endpoint, options, headers, method, code, payload)
 INSERT INTO service(id, short_name, long_name, description, endpoint, options, headers, method, code, payload) VALUES ('f33f9ce5-e402-4823-8847-f380d1b7789b', 'ornl_service', 'ORNL On-Prem Service', 'ORNL Service used for syncing data between EDPub and on-prem systems', 'https://postman-echo.com/post', '{}', '{"Authorization": "ornl_service_authorization_secret"}', 'POST', 200, false);
@@ -48,6 +49,10 @@ INSERT INTO section VALUES ('36d021d3-c527-419b-9b77-a48d535c12a8', '19025579-99
 INSERT INTO section VALUES ('f1a2136a-7245-4462-a175-40164db59c7b', '19025579-99ca-4344-8610-704dae626343', 'ORNL DAAC Additional Questions', 15, '[]', '[]', '15df4fda-ed0d-417f-9124-558fb5e5b561');
 INSERT INTO section VALUES ('0f6771e5-a6e4-4f19-ac85-cc48f526c146', '19025579-99ca-4344-8610-704dae626343', 'PO.DAAC Additional Questions', 16, '[]', '[]', '6b3ea184-57c5-4fc5-a91b-e49708f91b67');
 INSERT INTO section VALUES ('e9675ba5-e494-4275-8778-cd3fa68371e3', '19025579-99ca-4344-8610-704dae626343', 'SEDAC Additional Questions', 17, '[]', '[]', '00dcf32a-a4e2-4e55-a0d1-3a74cf100ca1');
+INSERT INTO section VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b51', '19025579-99ca-4344-8611-704dae626343', 'Contact Information', 0, '[]', '[]', NULL);
+INSERT INTO section VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b52', '19025579-99ca-4344-8611-704dae626343', 'Funding Information', 1, '[]', '[]', NULL);
+INSERT INTO section VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b53', '19025579-99ca-4344-8611-704dae626343', 'General Information', 2, '[]', '[]', NULL);
+INSERT INTO section VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b54', '19025579-99ca-4344-8611-704dae626343', 'Technical Information', 3, '[]', '[]', NULL);
 
 -- Question(id, short_name, version, long_name, text, help, required, created_at)
 INSERT INTO question VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e2', 'data_producer_info', 1, 'Primary Data Producer', 'Who is the primary person responsible for the collection or creation of this data product?', 'The Primary Data Producer is often the Principal Investigator, Project Scientist, or Project Manager.');
@@ -340,6 +345,7 @@ INSERT INTO privilege VALUES ('REQUEST_INITIALIZE');
 INSERT INTO privilege VALUES ('REQUEST_SUBMIT');
 INSERT INTO privilege VALUES ('REQUEST_REVIEW');
 INSERT INTO privilege VALUES ('REQUEST_REVIEW_MANAGER');
+INSERT INTO privilege VALUES ('REQUEST_REVIEW_ESDIS');
 INSERT INTO privilege VALUES ('REQUEST_REASSIGN');
 INSERT INTO privilege VALUES ('REQUEST_ADDUSER');
 INSERT INTO privilege VALUES ('REQUEST_REMOVEUSER');
@@ -372,6 +378,7 @@ INSERT INTO edprole_privilege VALUES ('19ac227b-e96c-46fa-a378-cf82c461b669', 'R
 INSERT INTO edprole_privilege VALUES ('19ac227b-e96c-46fa-a378-cf82c461b669', 'REQUEST_REVIEW');
 INSERT INTO edprole_privilege VALUES ('19ac227b-e96c-46fa-a378-cf82c461b669', 'NOTE_REPLY');
 INSERT INTO edprole_privilege VALUES ('19ac227b-e96c-46fa-a378-cf82c461b669', 'NOTE_ADDUSER');
+INSERT INTO edprole_privilege VALUES ('19ac227b-e96c-46fa-a378-cf82c461b669', 'USER_READ');
 
 -- RolePrivilege(edprole_id, privilege) Data Producer
 INSERT INTO edprole_privilege VALUES ('804b335c-f191-4d26-9b98-1ec1cb62b97d', 'REQUEST_INITIALIZE');
@@ -395,6 +402,7 @@ INSERT INTO edprole_privilege VALUES ('a5b4947a-67d2-434e-9889-59c2fad39676', 'G
 INSERT INTO edprole_privilege VALUES ('a5b4947a-67d2-434e-9889-59c2fad39676', 'CREATE_STEPREVIEW');
 INSERT INTO edprole_privilege VALUES ('a5b4947a-67d2-434e-9889-59c2fad39676', 'REMOVE_STEPREVIEW');
 INSERT INTO edprole_privilege VALUES ('a5b4947a-67d2-434e-9889-59c2fad39676', 'METRICS_READ');
+INSERT INTO edprole_privilege VALUES ('a5b4947a-67d2-434e-9889-59c2fad39676', 'USER_READ');
 
 --RolePrivilege(edprole_id, privilege) DAAC Data Manager
 INSERT INTO edprole_privilege VALUES ('2aa89c57-85f1-4611-812d-b6760bb6295c', 'REQUEST_DAACREAD');
@@ -411,6 +419,8 @@ INSERT INTO edprole_privilege VALUES ('2aa89c57-85f1-4611-812d-b6760bb6295c', 'C
 INSERT INTO edprole_privilege VALUES ('2aa89c57-85f1-4611-812d-b6760bb6295c', 'REMOVE_STEPREVIEW');
 INSERT INTO edprole_privilege VALUES ('2aa89c57-85f1-4611-812d-b6760bb6295c', 'NOTE_REPLY');
 INSERT INTO edprole_privilege VALUES ('2aa89c57-85f1-4611-812d-b6760bb6295c', 'METRICS_READ');
+INSERT INTO edprole_privilege VALUES ('2aa89c57-85f1-4611-812d-b6760bb6295c', 'NOTE_ADDUSER');
+INSERT INTO edprole_privilege VALUES ('2aa89c57-85f1-4611-812d-b6760bb6295c', 'NOTE_REMOVEUSER');
 
 --RolePrivilege(edprole_id, privilege) Observer
 INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'REQUEST_DAACREAD');
@@ -423,6 +433,10 @@ INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'N
 INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'NOTE_REMOVEUSER');
 INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'REQUEST_ASSIGNDAAC');
 INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'METRICS_READ');
+INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'REQUEST_REVIEW_ESDIS');
+INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'CREATE_STEPREVIEW');
+INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'REMOVE_STEPREVIEW');
+INSERT INTO edprole_privilege VALUES ('4be6ca4d-6362-478b-8478-487a668314b1', 'USER_READ');
 
 -- UserRole(edpuser_id, edprole_id)
 INSERT INTO edpuser_edprole VALUES ('1b10a09d-d342-4eee-a9eb-c99acd2dde17', '75605ac9-bf65-4dec-8458-93e018dcca97');
@@ -464,14 +478,22 @@ INSERT INTO step(step_name, type, data) VALUES ('map_from_mmt', 'action', '{"rol
 INSERT INTO step(step_name, type, data) VALUES ('publish_to_cmr', 'action', '{"rollback":"map_from_mmt","type": "action"}');
 INSERT INTO step(step_name, type, action_id) VALUES ('send_to_mmt', 'action', '3fe93672-cd91-45d4-863b-c6d0d63f8c8c');
 INSERT INTO step (step_name, type, data) VALUES ('cost_model', 'upload', '{"rollback":"data_accession_request_form_review","type": "review"}');
-INSERT INTO step(step_name, type, data) VALUES ('daac_assignment', 'action', '{"rollback":"data_accession_request_form_review","type": "review"}');
+INSERT INTO step(step_name, type, data) VALUES ('daac_assignment', 'action', '{"rollback":"additional_review_question","type": "action"}');
+INSERT INTO step(step_name, step_status_label, type, data) VALUES ('daac_assignment_final', 'Final DAAC Assignment', 'action', '{"rollback":"esdis_final_review","type": "review"}');
+INSERT INTO step(step_name, step_status_label, type, data) VALUES ('esdis_final_review', 'ESDIS Final Review', 'review', '{"rollback":"data_accession_request_form_review","type": "review", "form_id":"6c544723-241c-4896-a38c-adbc0a364293"}');
+INSERT INTO step(step_name, type, form_id) VALUES ('assignment_form_data_accession_request_form', 'form', '19025579-99ca-4344-8611-704dae626343');
+INSERT INTO step(step_name, type, data) VALUES ('additional_review_question', 'action', '{"rollback":"assignment_form_data_accession_request_form","type": "form","form":"19025579-99ca-4344-8611-704dae626343"}');
 
--- Unknown DAAC
+-- Accession Workflow
 -- StepEdge(workflow_id, step_name, next_step_name)
-INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'init', 'data_accession_request_form');
+INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'init', 'assignment_form_data_accession_request_form');
+INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'assignment_form_data_accession_request_form', 'additional_review_question');
+INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'additional_review_question', 'daac_assignment');
+INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'daac_assignment', 'data_accession_request_form');
 INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'data_accession_request_form', 'data_accession_request_form_review');
-INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'data_accession_request_form_review', 'daac_assignment');
-INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'daac_assignment', 'close');
+INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'data_accession_request_form_review', 'esdis_final_review');
+INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'esdis_final_review', 'daac_assignment_final');
+INSERT INTO step_edge VALUES ('3335970e-8a9b-481b-85b7-dfaaa3f5dbd9', 'daac_assignment_final', 'close');
 
 
 -- GHRC
@@ -618,3 +640,93 @@ INSERT INTO daac VALUES ('cdccdd71-cbe2-4220-8569-a6355ea24f3f', 'Example', 'Exa
 
 -- upload_step(id, step_name, upload_destination, category_type, help_text)
 INSERT INTO upload_step (step_name, upload_destination, category_type, help_text) VALUES ('cost_model', 'DAR_Uploads', 'cost_model', 'Please provide a cost model file. Files must be less than 5 GB and cannot include .exe or .dll extensions.');
+
+-- Data Assignment Form EDPUB-1435
+-- Question(id, short_name, version, long_name, text, help, required, created_at)
+INSERT INTO question VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e3', 'assignment_form_principal_investigator', 1, 'Principal Investigator', 'Who is the primary person responsible for managing the research that is creating the data you are hoping to publish at a NASA DAAC?', '');
+INSERT INTO question VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e4', 'assignment_form_data_submission_poc', 1, 'Data Submission Point of Contact', 'Who should ESDIS contact with questions regarding this data submission request?', 'This person should have in-depth knowledge of this project and the data it will be producing or has produced, allowing them to provide additional information as needed.');
+INSERT INTO question VALUES ('8a364184-42ac-48fe-b831-acb2eb08c730', 'assignment_form_funding_organization', 1, 'Funding Organization', 'What organization(s) funded the creation of this data?','',True);
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d731', 'assignment_form_funding_program', 1, 'Funding Program/Element', 'Under what program or program element within the funding organization was this data created?', 'Examples include NASA programs such as MEaSUREs, Terrestrial Hydrology, Earth Venture, ACCESS, or AIST.');
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d732', 'assignment_form_program_officer', 1, 'Program Officer', 'If known, provide the name of the NASA Headquarters Program Officer responsible for the program under which this project is funded.', '');
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d733', 'assignment_form_releated_projects', 1, 'Releated Projects', ' Is your data related to any NASA-funded project, mission, or campaign, or to any data products in NASA’s archive of Earth Science data?', '');
+
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d734', 'assignment_form_project_name', 1, 'Project Name', 'Please provide the full name of the project that produced this data. If there is an acronym or shortened project name that you use to refer to the project, please provide that in parentheses.', '');
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d735', 'assignment_form_project_desc', 1, 'Project Description', 'Please provide a brief description of this project.', 'The description should mimic a journal abstract and should provide a reader with the information needed to quickly understand the relevance and usefulness of your data products, including information about the spatial region(s) and temporal period(s) covered by your data.');
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d736', 'assignment_form_science_value', 1, 'Science Value', 'What is the science value of your data?', 'For example, describe the benefits to the science community, the strengths compared to similar data products and/or other data products that it complements.');
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d737', 'assignment_form_data_sub_req', 1, 'Reason for Data Submission Request', 'Why are you requesting to have your data archived and distributed at a DAAC?', 'For example, you have been instructed by a NASA program manager to archive your data at a DAAC, or you want your data to be distributed with related data products.');
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d738', 'assignment_form_data_sub_req_approval', 1, 'Dependencies for Data Submission Request  Approval', 'Do you have any dependencies related to this project’s data being approved for publication at a DAAC?', 'For example, you are publishing a paper and the publisher requires that your data be archived, or your project has a requirement to publish data by a certain time.');
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d739', 'assignment_form_open_data_policy', 1, 'Open Data Policy', ' Can your data be publicly released in compliance with NASA’s Open Data Policy?', 'For a description of the open data Policy, please refer to the <a href="https://www.earthdata.nasa.gov/engage/open-data-services-software-policies/data-information-guidance" target=_blank>NASA Earthdata Data and Information Policy <i class="fas fa-external-link-alt"></i></a>.');
+INSERT INTO question VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d740', 'assignment_form_project_documentation', 1, 'Project Documentation', 'Please upload the following document(s), if available: Program Level Requirements Appendix (PLRA), Data Management Plan (DMP), Project Proposal (i.e. the proposal you submitted in order to fund this project).', 'If there are other documents you would like us to consider, please upload those here as well. For example, these documents may include descriptions of your data, data quality, processing methods, or instruments. Files must be less than 5 GB and cannot include .exe or .dll extensions.');
+
+INSERT INTO question VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_form_data_format', 1, 'Data Format', 'What file format(s) does this data include?', 'For a list of NASA-approved data formats, please refer to the <a href="https://earthdata.nasa.gov/esdis/eso/standards-and-references#data-formats" target=_blank>NASA Earthdata Standards and Practices <i class="fas fa-external-link-alt"></i></a> web page.', True);
+INSERT INTO question VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d35', 'assignment_form_data_producers_table_info', 1, 'Data Products to be published', 'To the extent possible, please complete the table below for each data product generated by your project:', 'Data Product: A logically meaningful grouping of similar or related data which is packaged for distribution to users and represented with a single title and DOI.', True);
+
+-- SectionQuestion(section_id, question_id, list_order, required_if, show_if))
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b51', '80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e3', 0, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b51', '80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e4', 1, '[]', '[]');
+
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b52', '8a364184-42ac-48fe-b831-acb2eb08c730', 0, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b52', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d731', 1, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b52', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d732', 2, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b52', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d733', 3, '[]', '[]');
+
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b53', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d734', 0, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b53', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d735', 1, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b53', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d736', 2, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b53', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d737', 3, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b53', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d738', 4, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b53', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d739', 5, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b53', '4ecc885f-daf8-4bc6-a8cd-d30c2a54d740', 6, '[]', '[]');
+
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b54', '50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 0, '[]', '[]');
+INSERT INTO section_question VALUES ('1b4f110b-fea3-444f-b52c-c85008cf3b54', '50e8d566-b9ab-4bd9-9adc-92a3c8fb5d35', 1, '[]', '[]');
+
+-- Input(question_id, control_id, list_order, label, type, enums, attributes, required_if, show_if, required))
+INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e3', 'assignment_form_principal_investigator_fullname', 0, 'Full Name', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e3', 'assignment_form_principal_investigator_organization', 1, 'Organization', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e3', 'assignment_form_principal_investigator_email', 2, 'Email', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e3', 'assignment_form_principal_investigator_orcid', 3, 'ORCID', 'text', '{}', '{}', '[]','[]',  False);
+
+INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e4', 'assignment_form_data_submission_poc_name', 0, 'Full Name', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e4', 'assignment_form_data_submission_poc_organization', 1, 'Organization', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e4', 'assignment_form_data_submission_poc_email', 2, 'Email', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e4', 'assignment_form_data_submission_poc_orcid', 3, 'ORCID', 'text', '{}', '{}', '[]','[]',  False);
+
+INSERT INTO input VALUES ('8a364184-42ac-48fe-b831-acb2eb08c730', 'assignment_form_funding_nasa', 0, 'NASA', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('8a364184-42ac-48fe-b831-acb2eb08c730', 'assignment_form_funding_noaa', 1, 'NOAA', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('8a364184-42ac-48fe-b831-acb2eb08c730', 'assignment_form_funding_nsf', 2, 'NSF', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('8a364184-42ac-48fe-b831-acb2eb08c730', 'assignment_form_funding_usgs', 3, 'USGS', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('8a364184-42ac-48fe-b831-acb2eb08c730', 'assignment_form_funding_university', 4, 'University', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('8a364184-42ac-48fe-b831-acb2eb08c730', 'assignment_form_funding_other', 5, 'Other', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('8a364184-42ac-48fe-b831-acb2eb08c730', 'assignment_form_funding_organization_other', 6, 'If University or Other, please give the name here', 'text', '{}', '{}', '[{"field": "assignment_form_funding_other","value": "true"}, {"field": "assignment_form_funding_university","value": "true"}]','[]',  False);
+
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d731', 'assignment_form_funding_program_name', 0, '', 'text', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d732', 'assignment_form_program_officer_name', 0, '', 'text', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d733', 'assignment_form_releated_projects_yes_no', 0, '', 'radio', '["Yes","No"]', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d733', 'assignment_form_releated_projects_yes_explanation', 1, 'If Yes, please provide a brief explanation', 'text', '{}', '{}', '[{"field": "assignment_form_releated_projects_yes_no","value": "Yes"}]','[]',  False);
+
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d734', 'assignment_form_project_name_info', 0, '', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d735', 'assignment_form_project_desc_info', 0, '', 'textarea', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d736', 'assignment_form_science_value_info', 0, '', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d737', 'assignment_form_data_sub_req_info', 0, '', 'text', '{}', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d738', 'assignment_form_data_sub_req_approval_info', 0, '', 'radio', '["Yes","No"]', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d738', 'assignment_form_data_sub_req_approval_info_no', 1, 'If Yes, please provide a brief explanation.', 'text', '{}', '{}', '[{"field": "assignment_form_data_sub_req_approval_info","value": "Yes"}]','[]',  False);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d739', 'assignment_form_open_data_policy_info', 0, '', 'radio', '["Yes","No","Not Sure"]', '{}', '[]','[]',  True);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d739', 'assignment_form_open_data_policy_info_no', 1, 'If No or Not sure, please provide a brief explanation.', 'text', '{}', '{}', '[{"field": "assignment_form_open_data_policy_info","value": "No"}, {"field": "assignment_form_open_data_policy_info","value": "Not Sure"}]','[]',  False);
+
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d740', 'assignment_form_project_documentation', 0, '', 'file', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d740', 'assignment_form_project_documentation_url', 1, 'Alternatively, provide a URL to the document(s)', 'text', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('4ecc885f-daf8-4bc6-a8cd-d30c2a54d740', 'assignment_form_project_documentation_web', 2, 'Please provide a link to your project website, if available', 'text', '{}', '{}', '[]','[]',  False);
+
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_ascii', 0, 'ASCII', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_geotiff', 1, 'GeoTIFF', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_hdf5', 2, 'HDF 5', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_hdf_eos', 3, 'HDF-EOS 5', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_ogc_kml', 4, 'OGC KML', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_netcdf_4', 5, 'NetCDF-4', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_netcdf_classic', 6, 'NetCDF Classic', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_other', 7, 'Other', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_dont_know', 8, 'Don`t know yet', 'checkbox', '{}', '{}', '[]','[]',  False);
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d28', 'assignment_data_format_other_info', 9, 'If Other, please provide the data format(s)', 'text', '{}', '{}', '[{"field":"assignment_data_format_other","value":"true"}]','[]',  False);
+
+INSERT INTO input VALUES ('50e8d566-b9ab-4bd9-9adc-92a3c8fb5d35', 'assignment_form_data_producers_table', 0, '', 'table', '[{"key":"data_product_name","label":"Name of data product: How do you refer to the data product?","type":"text","editable":true},{"key": "data_prod_timeline","label": "Data Production Timeline: Include the start date and when do you expect data production to be complete.","type": "text","editable": true},{"key":"data_prod_volume","label":"Data Product Volume: What is the estimated or actual total volume of the data product upon completion of production?","type":"text","editable":true},{"key":"instrument_collect_data","label":"Instrument: What instrument  is used to collect data?","type":"text","editable":true}]', '{}', '[]','[]',  True);