+++ CHANGELOG.md
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix upload error for users not assigned to a group
 - Add default publication workflow
 - Update request details db query to remove daac long name from response
+- Added backend for handling ESDIS final review
 - Update notes to be associated with the submission's daac at time of creation
+- Updating backend to handle the needs_review case in daac selection
 
 ## 1.1.0